// src/type/model/column.d.ts
type ColumnRecord = {
    "updated": 1543830773,
    "description": "科学，以玩为本。",
    "topics": [
        {
            "url": "https://www.zhihu.com/api/v4/topics/20010094",
            "type": "topic",
            "id": "20010094",
            "name": "黑科技"
        },
        {
            "url": "https://www.zhihu.com/api/v4/topics/19555490",
            "type": "topic",
            "id": "19555490",
            "name": "手机游戏"
        }
    ],
    "can_manage": false,
    "intro": "公众号：狂丸科学",
    "is_following": false,
    "url_token": "kuangwan",
    "id": "kuangwan",
    "articles_count": 371,
    "accept_submission": false,
    "title": "狂丸科学",
    "url": "https://zhuanlan.zhihu.com/kuangwan",
    "comment_permission": "all",
    "created": 1497252346,
    "image_url": "https://pic2.zhimg.com/v2-59c79c22006310ae562e83988491fcbb_b.jpg",
    "author": {
        "is_followed": false,
        "avatar_url_template": "https://pic4.zhimg.com/v2-cba887679509294b7b929970e9038056_{size}.jpg",
        "uid": "856466662591987712",
        "user_type": "people",
        "is_following": false,
        "url_token": "kuang-wan-35-41",
        "id": "5f33ef832172f703f4434ca4e2b7a0b4",
        "description": "微信公众号：狂丸科学，获取科学学习资源",
        "name": "狂丸科学",
        "is_advertiser": false,
        "headline": "微信公众号：狂丸科学，获取科学学习资源",
        "gender": 1,
        "url": "/people/5f33ef832172f703f4434ca4e2b7a0b4",
        "avatar_url": "https://pic4.zhimg.com/v2-cba887679509294b7b929970e9038056_l.jpg",
        "is_org": false,
        "type": "people"
    },
    "followers": 18724,
    "type": "column"
}

export default ColumnRecord 